+++ src/Renderer/Repository/Issue/IssueFilter.ts
@@ -1,4 +1,4 @@
-import {GitHubQueryParser} from '../Infra/GitHubQueryParser';
+import {GitHubQueryParser} from '../../Infra/GitHubQueryParser';
 
 class _IssueFilter {
   // ---- filter ----
